fix: VBRuntimeCurrencyValue.Value integer division
StoredValue / ScaleFactor was long / int -> integer division, so a Currency
value read back as e.g. 12 instead of 12.3456. Cast the numerator to decimal.

+++ RDCore.SDK/Model/Values/Runtime/VBRuntimeCurrencyValue.cs
@@ -18,7 +18,7 @@ public readonly record struct VBRuntimeCurrencyValue : IRuntimeValue
     /// <summary>
     /// Gets the scaled decimal representation of the stored value.
     /// </summary>
-    public decimal Value => StoredValue / ScaleFactor;
+    public decimal Value => (decimal)StoredValue / ScaleFactor;
 
     /// <summary>
     /// Gets the scale factor.
